+++ src/main/java/com/github/blutorange/maven/plugin/closurecompiler/common/FileHelper.java
@@ -29,11 +29,10 @@ public class FileHelper {
                 absoluteFileToCanonicalFile(target.getAbsoluteFile()).toPath();
         if (base == null) {
             return targetPath.toString();
-        } else {
-            final var basePath =
-                    absoluteFileToCanonicalFile(base.getAbsoluteFile()).toPath();
-            return basePath.relativize(targetPath).toString();
         }
+
+        final var basePath = absoluteFileToCanonicalFile(base.getAbsoluteFile()).toPath();
+        return basePath.relativize(targetPath).toString().replace(File.separatorChar, '/');
     }
 
     /**
@@ -45,10 +44,10 @@ public class FileHelper {
         final var targetPath = target.toPath();
         if (base == null) {
             return targetPath.toString();
-        } else {
-            final var basePath = base.toPath();
-            return basePath.relativize(targetPath).toString();
         }
+
+        final var basePath = base.toPath();
+        return basePath.relativize(targetPath).toString().replace(File.separatorChar, '/');
     }
 
     /** Sames as {@link FileUtils#getFile(File, String...)}, but ignores empty names. */