+++ src/test/java/com/github/blutorange/maven/plugin/closurecompiler/test/MinifyMojoTest.java
@@ -1,45 +1,38 @@
 package com.github.blutorange.maven.plugin.closurecompiler.test;
 
-import static com.github.blutorange.maven.plugin.closurecompiler.common.FileHelper.*;
+import static com.github.blutorange.maven.plugin.closurecompiler.common.FileHelper.relativizePath;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.function.Function.identity;
-import static java.util.stream.Collectors.toList;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.junit.jupiter.api.Assertions.fail;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatNoException;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.github.blutorange.maven.plugin.closurecompiler.common.FileHelper;
 import io.takari.maven.testing.TestResources5;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.PrintStream;
 import java.nio.charset.Charset;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.logging.Logger;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import org.apache.commons.collections4.CollectionUtils;
-import org.apache.commons.collections4.SetUtils;
+import java.util.zip.GZIPInputStream;
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.maven.cli.MavenCli;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class MinifyMojoTest {
-
-    protected interface Action {
-        void run() throws Throwable;
-    }
 
     private static class EncodingProvider {
         private final File basedir;
@@ -48,15 +41,17 @@ public class MinifyMojoTest {
         EncodingProvider(File basedir) throws IOException {
             this.basedir = basedir;
             var encodingData = new File(basedir, "encoding.txt");
-            if (encodingData.exists()) {
-                var encodingLines = FileUtils.readLines(encodingData, UTF_8);
-                for (var encodingLine : encodingLines) {
-                    var parts = encodingLine.split("=");
-                    if (parts.length == 2) {
-                        var relativePath = parts[0];
-                        var encoding = Charset.forName(parts[1]);
-                        encodingMap.put(relativePath, encoding);
-                    }
+            if (!encodingData.exists()) {
+                return;
+            }
+
+            var encodingLines = FileUtils.readLines(encodingData, UTF_8);
+            for (var encodingLine : encodingLines) {
+                var parts = encodingLine.split("=");
+                if (parts.length == 2) {
+                    var relativePath = parts[0];
+                    var encoding = Charset.forName(parts[1]);
+                    encodingMap.put(relativePath, encoding);
                 }
             }
         }
@@ -68,84 +63,61 @@ public class MinifyMojoTest {
     }
 
     private static class MavenResult {
-        final String errString;
         final String outString;
 
-        public MavenResult(String outString, String errString) {
+        public MavenResult(String outString) {
             this.outString = outString;
-            this.errString = errString;
         }
 
-        public String getErrString() {
-            return errString;
-        }
-
+        /**
+         * The combined output from stdout and stderr.
+         *
+         * @return The combined output from stdout and stderr.
+         */
         public String getOutString() {
             return outString;
         }
     }
 
-    private final Logger LOG = Logger.getLogger(MinifyMojoTest.class.getCanonicalName());
+    private final Logger LOG = LoggerFactory.getLogger(MinifyMojoTest.class.getCanonicalName());
 
     @RegisterExtension
     final TestResources5 testResources = new TestResources5("src/test/resources/projects", "target/test-projects");
 
     private void assertDirContent(File basedir) throws IOException {
-        var expected = new File(basedir, "expected");
         var actual = new File(new File(basedir, "target"), "test");
-        var expectedFiles = expected.exists() ? listFiles(expected) : new HashMap<String, File>();
-        var actualFiles = actual.exists() ? listFiles(actual) : new HashMap<String, File>();
-        LOG.info("Comparing actual files [\n"
-                + actualFiles.values().stream().map(File::getAbsolutePath).collect(Collectors.joining(",\n")) + "\n]");
-        LOG.info("to the expected files [\n"
-                + expectedFiles.values().stream().map(File::getAbsolutePath).collect(Collectors.joining(",\n"))
-                + "\n]");
-        assertFalse(
-                expectedFiles.isEmpty(),
-                "There must be at least one expected file. Add a file 'nofiles' if you expect there to be no files");
-        var encodingProvider = new EncodingProvider(basedir);
+        var expected = new File(basedir, "expected");
+        assertThat(expected.exists())
+                .as("Expected directory %s must exist", expected)
+                .isTrue();
+
+        var expectedFiles = listFiles(expected);
+        assertThat(expectedFiles)
+                .as("There must be at least one expected file. Add a file 'nofiles' if you expect there to be no files")
+                .isNotEmpty();
+
         if (expectedFiles.size() == 1
                 && "nofiles".equals(expectedFiles.values().iterator().next().getName())) {
-            // Expect there to be no output files
-            assertEquals(0, actualFiles.size());
-        } else {
-            assertEquals(
-                    expectedFiles.size(),
-                    actualFiles.size(),
-                    "Number of expected files must match the number of produced files. "
-                            + diffFiles(basedir, expectedFiles, actualFiles));
-            assertTrue(
-                    CollectionUtils.isEqualCollection(expectedFiles.keySet(), actualFiles.keySet()),
-                    "Expected file names must match the produced file names. "
-                            + diffFiles(basedir, expectedFiles, actualFiles));
-            expectedFiles.forEach((key, expectedFile) -> {
-                var actualFile = actualFiles.get(key);
-                try {
-                    compareFiles(expectedFile, actualFile, encodingProvider);
-                } catch (IOException e) {
-                    throw new RuntimeException(e);
-                }
-            });
+            if (actual.exists()) {
+                assertThat(listFiles(actual))
+                        .as("Expecting no output files in %s", actual)
+                        .isEmpty();
+            }
+            return;
         }
-    }
 
-    private static String diffFiles(File basedir, Map<String, File> expectedFiles, Map<String, File> actualFiles) {
-        final var testDir = new File(basedir, "target/test");
-        final var expectedButNotPresent = SetUtils.difference(expectedFiles.keySet(), actualFiles.keySet()).stream()
-                .map(x -> new File(testDir, x))
-                .map(FileHelper::absoluteFileToCanonicalFile)
-                .collect(toList());
-        final var presentButNotExpected = SetUtils.difference(actualFiles.keySet(), expectedFiles.keySet()).stream()
-                .map(actualFiles::get)
-                .collect(toList());
-        final var messages = new ArrayList<String>();
-        if (!expectedButNotPresent.isEmpty()) {
-            messages.add("Expected files that are not present: <" + expectedButNotPresent + ">.");
-        }
-        if (!presentButNotExpected.isEmpty()) {
-            messages.add("Present files that were not expected to be present: <" + presentButNotExpected + ">");
-        }
-        return String.join(" ", messages);
+        assertThat(actual.exists()).as("Actual directory %s must exist", actual).isTrue();
+
+        var actualFiles = listFiles(actual);
+        assertThat(actualFiles).containsOnlyKeys(expectedFiles.keySet());
+
+        var encodingProvider = new EncodingProvider(basedir);
+        expectedFiles.forEach((key, expectedFile) -> {
+            var actualFile = actualFiles.get(key);
+            assertThatNoException()
+                    .isThrownBy(() ->
+                            compareFiles(expectedFile, actualFile, encodingProvider.determineEncoding(expectedFile)));
+        });
     }
 
     private void clean(File basedir) throws IOException {
@@ -153,51 +125,50 @@ public class MinifyMojoTest {
         if (target.exists()) {
             FileUtils.forceDelete(target);
         }
-        assertFalse(target.exists());
+        assertThat(target.exists())
+                .as("Target folder must not exist anymore: %s", target)
+                .isFalse();
     }
 
-    private void compareFiles(File expectedFile, File actualFile, EncodingProvider encodingProvider)
-            throws IOException {
-        var encoding = encodingProvider.determineEncoding(expectedFile);
-        var expectedLines = FileUtils.readLines(expectedFile, encoding);
-        var actualLines = FileUtils.readLines(actualFile, encoding);
-        assertTrue(
-                expectedFile.exists(),
-                "File with expected content does not exist: '" + actualFile.getAbsolutePath() + "'");
-        assertTrue(
-                actualFile.exists(),
-                "File with produced content does not exist: '" + actualFile.getAbsolutePath() + "'");
+    private void compareFiles(File expectedFile, File actualFile, Charset charset) throws IOException {
+        final List<String> expectedLines;
+        final List<String> actualLines;
+
+        assertThat(expectedFile).exists();
+        assertThat(actualFile).exists();
+
+        if (expectedFile.getAbsolutePath().endsWith(".gz")) {
+            expectedLines = readLinesFromGzipFile(expectedFile, charset);
+            actualLines = readLinesFromGzipFile(actualFile, charset);
+        } else {
+            expectedLines = FileUtils.readLines(expectedFile, charset);
+            actualLines = FileUtils.readLines(actualFile, charset);
+        }
+
         // Ignore empty lines
         expectedLines.removeIf(StringUtils::isBlank);
         actualLines.removeIf(StringUtils::isBlank);
+
         // Check file contents
-        assertFalse(
-                expectedLines.isEmpty(),
-                "Expected file must contain at least one non-empty line: '" + actualFile.getAbsolutePath() + "'");
-        assertEquals(
-                expectedLines.size(),
-                actualLines.size(),
-                "Number of non-empty lines in expected file must match the generated number of lines: '"
-                        + actualFile.getAbsolutePath() + "'");
-        for (int i = 0, j = expectedLines.size(); i < j; ++i) {
-            assertEquals(
-                    expectedLines.get(i).trim(),
-                    actualLines.get(i).trim(),
-                    "Actual content of file '" + actualFile.getAbsolutePath() + "' differs from the expected content");
+        assertThat(expectedLines).as("Content of %s", expectedFile).isNotEmpty();
+        assertThat(actualLines)
+                .as("Content of %s has same number of non-empty lines as %s", actualFile, expectedFile)
+                .hasSameSizeAs(expectedLines);
+
+        var lines = expectedLines.size();
+        for (int i = 0; i < lines; ++i) {
+            assertThat(actualLines.get(i))
+                    .as("Line from %s should be the same as line from %s", actualFile, expectedFile)
+                    .isEqualToIgnoringNewLines(expectedLines.get(i));
         }
     }
 
-    private <T extends Throwable> void expectError(Class<T> error, Action runnable) {
-        try {
-            runnable.run();
-        } catch (Throwable e) {
-            if (error.isInstance(e)) {
-                return;
-            }
-            fail("Action threw an error of type " + e.getClass().getSimpleName()
-                    + ", but it is not of the expected type " + error.getSimpleName());
+    private List<String> readLinesFromGzipFile(File file, Charset charset) throws IOException {
+        try (var fileInputStream = new FileInputStream(file);
+                var gzipInputStream = new GZIPInputStream(fileInputStream)) {
+
+            return IOUtils.readLines(gzipInputStream, charset);
         }
-        fail("Action did not throw the expected error type " + error.getSimpleName());
     }
 
     private MavenResult invokeMaven(File pom, String goal, Collection<String> profiles) throws IOException {
@@ -207,15 +178,13 @@ public class MinifyMojoTest {
         args.add("-DskipTests");
         profiles.stream().flatMap(profile -> Stream.of("-P", profile)).forEach(args::add);
         System.setProperty("maven.multiModuleProjectDirectory", pom.getParent());
-        LOG.info("Invoking maven: " + StringUtils.join(args, " "));
-        try (final var out = new ByteArrayOutputStream();
-                final var err = new ByteArrayOutputStream()) {
-            try (final var outStream = new ChainedPrintStream(new PrintStream(out), System.out);
-                    final var errStream = new ChainedPrintStream(new PrintStream(err), System.err)) {
-                final var cli = new MavenCli();
-                cli.doMain(args.toArray(new String[0]), pom.getParent(), outStream, errStream);
-                return new MavenResult(out.toString(UTF_8), err.toString(UTF_8));
-            }
+        LOG.info("Invoking maven: {}", StringUtils.join(args, " "));
+        try (final var out = new ByteArrayOutputStream()) {
+            // We don't want to close ChainedPrintStream as it would close System.out, so we don't use
+            // try-with-resources
+            final var outStream = new ChainedPrintStream(new PrintStream(out), System.out);
+            new MavenCli().doMain(args.toArray(new String[0]), pom.getParent(), outStream, outStream);
+            return new MavenResult(out.toString(UTF_8));
         }
     }
 
@@ -224,28 +193,28 @@ public class MinifyMojoTest {
                 .collect(Collectors.toMap(file -> relativizePath(basedir, file), identity()));
     }
 
-    private MavenResult runMinify(String projectName) throws Exception {
-        return runMinify(projectName, new HashSet<>());
-    }
-
     private MavenResult runMinify(String projectName, Collection<String> profiles) throws Exception {
         final var parentDir = testResources.getBasedir("parent").getCanonicalFile();
         final var parentPom = new File(parentDir, "pom.xml");
         final var parentPomNew = new File(parentDir.getParentFile(), "pom.xml");
-        assertTrue(parentPom.exists());
+        assertThat(parentPom).exists();
         FileUtils.copyFile(parentPom, parentPomNew);
 
         final var basedir = testResources.getBasedir(projectName).getCanonicalFile();
         final var pom = new File(basedir, "pom.xml");
-        assertTrue(pom.exists());
+        assertThat(pom).exists();
 
         clean(basedir);
-        invokeMaven(parentPomNew, "install", Collections.emptySet());
+        invokeMaven(parentPomNew, "install", List.of());
         return invokeMaven(pom, "package", profiles);
     }
 
     private void runMinifyAndAssertDirContent(String projectName) throws Exception {
         runMinifyAndAssertDirContent(projectName, new HashSet<>());
+    }
+
+    private List<String> profiles(String... profiles) {
+        return List.of(profiles);
     }
 
     private void runMinifyAndAssertDirContent(String projectName, Collection<String> profiles) throws Exception {
@@ -292,10 +261,11 @@ public class MinifyMojoTest {
     @Test
     public void testExterns() throws Exception {
         // No externs declared, variable cannot be found, so minification should fail
-        expectError(AssertionError.class, () -> runMinifyAndAssertDirContent("externs", List.of("without-externs")));
+        assertThatThrownBy(() -> runMinifyAndAssertDirContent("externs", profiles("without-externs")))
+                .isInstanceOf(AssertionError.class);
 
         // Externs declared, variable can be found, so minification should succeed
-        runMinifyAndAssertDirContent("externs", List.of("createOlderFile", "with-externs"));
+        runMinifyAndAssertDirContent("externs", profiles("createOlderFile", "with-externs"));
     }
 
     @Test
@@ -330,9 +300,8 @@ public class MinifyMojoTest {
 
     @Test
     public void testOverwriteInputFilesDisabled() throws Exception {
-        MavenResult result = runMinify("overwriteInputFilesDisabled");
-        assertTrue(result.getErrString()
-                .contains("The source file [fileC.js] has the same name as the output file [fileC.js]"));
+        assertThat(runMinify("overwriteInputFilesDisabled", profiles()).getOutString())
+                .contains("The source file [fileC.js] has the same name as the output file [fileC.js]");
     }
 
     @Test
@@ -372,31 +341,30 @@ public class MinifyMojoTest {
 
     @Test
     public void testSkipIfExists() throws Exception {
-        // Output file does not exist, minification should run
-        expectError(AssertionError.class, () -> runMinifyAndAssertDirContent("skipif", List.of("skipIfExists")));
+        runMinify("skipif", List.of("skipIfExists"));
 
         // This creates the (older) output file, so the minification process should not run
-        runMinifyAndAssertDirContent("skipif", List.of("createOlderFile", "skipIfExists"));
+        runMinifyAndAssertDirContent("skipif", profiles("createOlderFile", "skipIfExists"));
 
         // Now force is enabled, minification should run
-        expectError(
-                AssertionError.class,
-                () -> runMinifyAndAssertDirContent(
-                        "skipif", Arrays.asList("createOlderFile", "skipIfExists", "force")));
+        assertThatThrownBy(() ->
+                        runMinifyAndAssertDirContent("skipif", profiles("createOlderFile", "skipIfExists", "force")))
+                .isInstanceOf(AssertionError.class);
     }
 
     @Test
     public void testSkipIfNewer() throws Exception {
         // Output file does not exist, minification should run
-        expectError(AssertionError.class, () -> runMinifyAndAssertDirContent("skipif", List.of("skipIfNewer")));
+        assertThatThrownBy(() -> runMinifyAndAssertDirContent("skipif", profiles("skipIfNewer")))
+                .isInstanceOf(AssertionError.class);
 
         // This creates the newer output file, so the minification process should not run
-        runMinifyAndAssertDirContent("skipif", Arrays.asList("createNewerFile", "skipIfNewer"));
+        runMinifyAndAssertDirContent("skipif", profiles("createNewerFile", "skipIfNewer"));
 
         // Now force is enabled, minification should run
-        expectError(
-                AssertionError.class,
-                () -> runMinifyAndAssertDirContent("skipif", Arrays.asList("createNewerFile", "skipIfNewer", "force")));
+        assertThatThrownBy(() ->
+                        runMinifyAndAssertDirContent("skipif", profiles("createNewerFile", "skipIfNewer", "force")))
+                .isInstanceOf(AssertionError.class);
     }
 
     @Test
@@ -422,5 +390,10 @@ public class MinifyMojoTest {
     @Test
     public void testUseTypesForOptimization() throws Exception {
         runMinifyAndAssertDirContent("usetypesforoptimization");
+    }
+
+    @Test
+    public void testGzipCompression() throws Exception {
+        runMinifyAndAssertDirContent("gzip-compression");
     }
 }