+++ src/test/resources/projects/gzip-compression/expected/nomerge_minify/a.js
@@ -1,0 +1,1 @@
+'use strict';console.log("Logging 1 (one)");console.log("Logging 1 (one)");console.log("Logging 1 (one)");