+++ src/test/resources/projects/gzip-compression/expected/nomerge_minify/b.js
@@ -1,0 +1,1 @@
+'use strict';console.log("Logging 2 (two)");console.log("Logging 2 (two)");console.log("Logging 2 (two)");