+++ src/test/resources/projects/gzip-compression/expected/nomerge_nominify/a.js
@@ -1,0 +1,3 @@
+console.log("Logging 1 (one)");
+console.log("Logging 1 (one)");
+console.log("Logging 1 (one)");