+++ src/test/resources/projects/gzip-compression/expected/nomerge_nominify/b.js
@@ -1,0 +1,3 @@
+console.log("Logging 2 (two)");
+console.log("Logging 2 (two)");
+console.log("Logging 2 (two)");