+++ src/test/resources/projects/outputfilename/expected/no-merge-no-minify-hello3.js
@@ -2,5 +2,5 @@ function main3() {
 	function sayHello3(name) {
 		console.log("Hello, " + name + "!");
 	}
-	sayHello3("blutorange3");
+	sayHello3("blutorange3");	
 } // comment 3